requirements.txt
haystack-ai
ollama-haystack
elasticsearch-haystack
rich
flask
flask_limiter
flask_cors
python-dotenv
bs4
trafilatura
aiohttp
